moved buses client to /internal/buses
