+++ internal/buses/cyprusbus/data/routes.txt
