+++ internal/buses/cyprusbus/integration_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/yanakipre/bot/internal/testtooling"
 
-	"github.com/yanakipre/bot/app/telegramsearch/internal/pkg/client/buses"
+	"github.com/yanakipre/bot/internal/buses"
 )
 
 func TestFullWorkflow(t *testing.T) {
