+++ internal/buses/cyprusbus/transport_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/yanakipre/bot/app/telegramsearch/internal/pkg/client/buses"
+	"github.com/yanakipre/bot/internal/buses"
 )
 
 type MockFetcher struct {
